Externalize local-ca-* arguments to be passed from the commandline and add check for files existence

+++ scripts/apiml_cm.sh
@@ -10,68 +10,77 @@ function usage {
     echo "     - new-service - adds new service"
     echo ""
     echo "  apiml_cm.sh --action setup"
-    echo "  apiml_cm.sh --action setup --local-ca-alias ... --local-ca-filename ... --local-ca-dname ... --local-ca-pass ... --local-ca-validity ..."
+    echo "  apiml_cm.sh --action setup --local-ca-alias ... --local-ca-filename ... --local-ca-dname ... --local-ca-password ... --local-ca-validity ..."
 }
 
 LOCAL_CA_ALIAS="localca"
 LOCAL_CA_FILENAME="keystore/local_ca/localca"
 LOCAL_CA_DNAME="CN=Zowe Development Instances Certificate Authority, OU=API Mediation Layer, O=Zowe Sample, L=Prague, S=Prague, C=CZ"
-LOCAL_CA_PASS="local_ca_password"
+LOCAL_CA_PASSWORD="local_ca_password"
 LOCAL_CA_VALIDITY=3650
 
 SERVER_ALIAS="localhost"
-SERVER_PASS="password"
+SERVER_PASSWORD="password"
 SERVER_FILENAME="keystore/localhost/localhost"
 SERVER_DNAME="CN=Zowe Service, OU=API Mediation Layer, O=Zowe Sample, L=Prague, S=Prague, C=CZ"
 SERVER_EXT="SAN=dns:localhost.localdomain,dns:localhost"
 SERVER_VALIDITY=3650
 
 function clean_local_ca {
-    rm -f ${LOCAL_CA_FILENAME}.keystore.p12 ${LOCAL_CA_FILENAME}.cer
+    if [[ -e ${LOCAL_CA_FILENAME}.keystore.p12 && -e ${LOCAL_CA_FILENAME}.cer ]];
+    then
+        rm -f ${LOCAL_CA_FILENAME}.keystore.p12 ${LOCAL_CA_FILENAME}.cer
+    fi
 }
 
 function clean_service {
-    rm -f ${SERVER_FILENAME}.keystore.p12 ${SERVER_FILENAME}.csr ${SERVER_FILENAME}_signed.cer ${SERVER_FILENAME}.truststore.p12
+    if [[ -e ${SERVER_FILENAME}.keystore.p12 && -e ${SERVER_FILENAME}.csr && -e ${SERVER_FILENAME}.cer && -e ${SERVER_FILENAME}.truststore.p12 ]];
+    then
+        rm -f ${SERVER_FILENAME}.keystore.p12 ${SERVER_FILENAME}.csr ${SERVER_FILENAME}_signed.cer ${SERVER_FILENAME}.truststore.p12
+    fi
 }
 
 function create_certificate_authority {
     echo "Generate keystore with the local CA private key and local CA public certificate:"
     keytool -genkeypair -v -alias ${LOCAL_CA_ALIAS} -keyalg RSA -keysize 2048 -keystore ${LOCAL_CA_FILENAME}.keystore.p12 \
-        -dname "${LOCAL_CA_DNAME}" -keypass ${LOCAL_CA_PASS} -storepass ${LOCAL_CA_PASS} -storetype PKCS12 -validity ${LOCAL_CA_VALIDITY} \
+        -dname "${LOCAL_CA_DNAME}" -keypass ${LOCAL_CA_PASSWORD} -storepass ${LOCAL_CA_PASSWORD} -storetype PKCS12 -validity ${LOCAL_CA_VALIDITY} \
         -ext KeyUsage="keyCertSign" -ext BasicConstraints:"critical=ca:true"
 
     echo "Export local CA public certificate:"
     keytool -export -v -alias ${LOCAL_CA_ALIAS} -file ${LOCAL_CA_FILENAME}.cer -keystore ${LOCAL_CA_FILENAME}.keystore.p12 -rfc \
-        -keypass ${LOCAL_CA_PASS} -storepass ${LOCAL_CA_PASS} -storetype PKCS12
+        -keypass ${LOCAL_CA_PASSWORD} -storepass ${LOCAL_CA_PASSWORD} -storetype PKCS12
 }
 
 function create_service_certificate_and_csr {
-    echo "Generate service private key and service:"
-    keytool -genkeypair -v -alias ${SERVER_ALIAS} -keyalg RSA -keysize 2048 -keystore ${SERVER_FILENAME}.keystore.p12 -keypass ${SERVER_PASS} -storepass ${SERVER_PASS} \
-        -storetype PKCS12 -dname "${SERVER_DNAME}" -validity ${SERVER_VALIDITY}
+    if [ ! -e "${SERVER_FILENAME}.keystore.p12" ];
+    then
+        echo "Generate service private key and service:"
+        keytool -genkeypair -v -alias ${SERVER_ALIAS} -keyalg RSA -keysize 2048 -keystore ${SERVER_FILENAME}.keystore.p12 -keypass ${SERVER_PASSWORD} -storepass ${SERVER_PASSWORD} \
+            -storetype PKCS12 -dname "${SERVER_DNAME}" -validity ${SERVER_VALIDITY}
 
-    echo "Generate CSR for the the service certificate:"
-    keytool -certreq -v -alias ${SERVER_ALIAS} -keystore ${SERVER_FILENAME}.keystore.p12 -storepass ${SERVER_PASS} -file ${SERVER_FILENAME}.csr \
-        -keyalg RSA -storetype PKCS12 -dname "${SERVER_DNAME}" -validity ${SERVER_VALIDITY}
+        echo "Generate CSR for the the service certificate:"
+        keytool -certreq -v -alias ${SERVER_ALIAS} -keystore ${SERVER_FILENAME}.keystore.p12 -storepass ${SERVER_PASSWORD} -file ${SERVER_FILENAME}.csr \
+            -keyalg RSA -storetype PKCS12 -dname "${SERVER_DNAME}" -validity ${SERVER_VALIDITY}
+    fi
 }
 
 function sign_csr_using_local_ca {
     echo "Sign the CSR using the Certificate Authority:"
     keytool -gencert -v -infile ${SERVER_FILENAME}.csr -outfile ${SERVER_FILENAME}_signed.cer -keystore ${LOCAL_CA_FILENAME}.keystore.p12 \
-        -alias ${LOCAL_CA_ALIAS} -keypass ${LOCAL_CA_PASS} -storepass ${LOCAL_CA_PASS} -storetype PKCS12 \
+        -alias ${LOCAL_CA_ALIAS} -keypass ${LOCAL_CA_PASSWORD} -storepass ${LOCAL_CA_PASSWORD} -storetype PKCS12 \
         -ext ${SERVER_EXT} -ext KeyUsage:critical=keyEncipherment,digitalSignature,nonRepudiation,dataEncipherment -ext ExtendedKeyUsage=clientAuth,serverAuth -rfc \
         -validity ${SERVER_VALIDITY}
 }
 
 function import_signed_certificate_and_ca_certificate {
     echo "Import the Certificate Authority to the truststore:"
-    keytool -importcert -v -trustcacerts -noprompt -file ${LOCAL_CA_FILENAME}.cer -alias ${LOCAL_CA_ALIAS} -keystore ${SERVER_FILENAME}.truststore.p12 -storepass ${SERVER_PASS} -storetype PKCS12
+    keytool -importcert -v -trustcacerts -noprompt -file ${LOCAL_CA_FILENAME}.cer -alias ${LOCAL_CA_ALIAS} -keystore ${SERVER_FILENAME}.truststore.p12 -storepass ${SERVER_PASSWORD} -storetype PKCS12
 
     echo "Import the Certificate Authority to the keystore:"
-    keytool -importcert -v -trustcacerts -noprompt -file ${LOCAL_CA_FILENAME}.cer -alias ${LOCAL_CA_ALIAS} -keystore ${SERVER_FILENAME}.keystore.p12 -storepass ${SERVER_PASS} -storetype PKCS12
+    keytool -importcert -v -trustcacerts -noprompt -file ${LOCAL_CA_FILENAME}.cer -alias ${LOCAL_CA_ALIAS} -keystore ${SERVER_FILENAME}.keystore.p12 -storepass ${SERVER_PASSWORD} -storetype PKCS12
 
     echo "Import the signed CSR to the keystore:"
-    keytool -importcert -v -trustcacerts -noprompt -file ${SERVER_FILENAME}_signed.cer -alias ${SERVER_ALIAS} -keystore ${SERVER_FILENAME}.keystore.p12 -storepass ${SERVER_PASS} -storetype PKCS12
+    keytool -importcert -v -trustcacerts -noprompt -file ${SERVER_FILENAME}_signed.cer -alias ${SERVER_ALIAS} -keystore ${SERVER_FILENAME}.keystore.p12 -storepass ${SERVER_PASSWORD} -storetype PKCS12
 }
 
 function setup_local_ca {
@@ -100,6 +109,46 @@ while [ "$1" != "" ]; do
                                 exit 1
     esac
     shift
+    case $1 in
+        --local-ca-alias )      shift
+                                LOCAL_CA_ALIAS=$1
+                                ;;
+        * )                     usage
+                                exit 1
+    esac
+    shift
+    case $1 in
+        --local-ca-filename )   shift
+                                LOCAL_CA_FILENAME=$1
+                                ;;
+        * )                     usage
+                                exit 1
+    esac
+    shift
+    case $1 in
+        --local-ca-dname )      shift
+                                LOCAL_CA_DNAME=$1
+                                ;;
+        * )                     usage
+                                exit 1
+    esac
+    shift
+    case $1 in
+        --local-ca-password )   shift
+                                LOCAL_CA_PASSWORD=$1
+                                ;;
+        * )                     usage
+                                exit 1
+    esac
+    shift
+    case $1 in
+        --local-ca-validity )   shift
+                                LOCAL_CA_VALIDITY=$1
+                                ;;
+        * )                     usage
+                                exit 1
+    esac
+    shift
 done
 
 case $ACTION in
@@ -114,3 +163,4 @@ case $ACTION in
         usage
         exit 1
 esac
+
